extract singletenancy from main
